requirements.txt
cryptography
fastapi
redis
python-jose
uvicorn
aiofiles
python-multipart